Meta files for one model

+++ src/engine_core/assets/src/HMeta.hpp
@@ -16,7 +16,7 @@ namespace Hush
 		static constexpr uint16_t VERSION = 1;
 
 		uint16_t version = VERSION;
-		uint32_t id = 0;
+		uint64_t id = 0;
 		uint64_t sourceHash = 0;
 		uint64_t sourceMTime = 0;
 		std::string assetType; // "texture" | "shader" | "model" | "unknown"
